Refactor: standardize User controller and service reponses and add new update method in repository

+++ src/main/java/org/example/service/UserService.java
@@ -1,8 +1,13 @@
 package org.example.service;
 
+import org.example.exception.UserAlreadyExistsException;
+import org.example.exception.UserNotFoundException;
 import org.example.model.User;
 import org.example.repository.UserRepository;
+
 import org.springframework.stereotype.Service;
+
+import java.util.List;
 
 @Service
 public class UserService {
@@ -13,20 +18,34 @@ public class UserService {
         this.userRepository = userRepository;
     }
 
+    public User createUser(User user) {
 
-    public String getAll() {
-        return userRepository.findAll();
-    }
-
-    public String createUser(User user) {
+        User existing = userRepository.findById(user.getId());
+        if (existing != null){
+            throw new UserAlreadyExistsException("User already exists");
+        }
         return userRepository.save(user);
     }
 
-    public String deleteUser(int id) {
-        return userRepository.deleteById(id);
+    public List<User> getAll() {
+        return userRepository.findAll();
+    }
+
+    public void deleteUser(int id) {
+        User user = userRepository.findById(id);
+        if (user == null){
+            throw new UserNotFoundException("User with such id does not exist");
+        }
+        userRepository.deleteById(id);
     }
 
     public User getById(int id) {
+        User user = userRepository.findById(id);
+
+        if (user == null) {
+            throw new UserNotFoundException("User with such id does not exist");
+        }
+
         return userRepository.findById(id);
     }
 
@@ -39,7 +58,7 @@ public class UserService {
 
         existingUser.setUsername(user.getUsername());
         existingUser.setPassword(user.getPassword());
-        userRepository.save(existingUser);
+        userRepository.update(existingUser);
         return existingUser;
     }
 
@@ -56,7 +75,7 @@ public class UserService {
             existingUser.setPassword(updateUser.getPassword());
         }
 
-        userRepository.save(existingUser);
+        userRepository.update(existingUser);
         return existingUser;
     }
 }
